make messages stream more readable
