+++ common/src/types.rs
@@ -16,8 +16,16 @@ impl InstallationId {
     }
 }
 
-#[derive(Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
+#[derive(Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
 pub struct GroupId(bytes::Bytes);
+
+impl std::fmt::Debug for GroupId {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        f.debug_tuple("GroupId")
+            .field(&crate::fmt::debug_hex(&self.0))
+            .finish()
+    }
+}
 
 impl AsRef<[u8]> for GroupId {
     fn as_ref(&self) -> &[u8] {
@@ -47,6 +55,12 @@ impl std::borrow::Borrow<[u8]> for GroupId {
 impl From<Vec<u8>> for GroupId {
     fn from(v: Vec<u8>) -> GroupId {
         GroupId(v.into())
+    }
+}
+
+impl From<&[u8]> for GroupId {
+    fn from(v: &[u8]) -> GroupId {
+        GroupId(v.to_vec().into())
     }
 }
